fix(Tooltip): Use native html to create tooltips
The Material-UI Tooltip component does not work properly when rendered in an
iframe, so hovering over a component with a tooltip causes the iframe to resize
quickly which makes the UI jump around. Since this leaves the GUI unusable and
since Tensorboard renders everything inside of an iframe we will use native
HTML for now.