feat(bootstrap): Set up play/pause button and debug slider
This is a commented version of the algorithm using bootstrap.
In this version the sliders are positioned on the bottom of the page, as they should.
Needs to be fixed: the handleSliderChange function is not working inside Form.Group. This is due to the fact that value is an undefined variable.
The Slider MUI component is left here as a comparison and debug feature.