feat(bootstrap): Include Badge, fix val, replace UseActor
Introduce a Badge to display the positions of the planes.
Use event _e to capture the value from target. Target is an element that emitted such event _e.
Convert the data string in handleSliderChange into a Number so we can use toFixed in Badge.
Replace the useActor with useSelector for the sake of rendering optimization.