feat(bootstrap): Align buttons and sliders, avoid memory leak, clean code
In this version we update the changes made in commit `c6aa754`.
We align the buttons and sliders by using the classes `input-group` and `input-group-prepend`.
We avoid a memory leak in useEffect by removing the `onResize` event listener.
We clean the code by deleting the comments, and unnecessary "string interpolations".
We delete the toFix because we do not have such precision and end up printing unnecessary zeros.

To be fixed: PlaneSliders can be optimized by avoiding unnecessary re-renders.