feat(PlaneSliders/Style): Shift sliders down when in 2D mode
Shift slider down when visualizing a cut (plane) of the figure.
This is done by using `display:none` so we get rid of the element in the DOM, which avoids the space that otherwise would be created when using `visibility:hidden`.

In PlaneSliders we also deleted mb-2 in the `className=input-group` - it makes no difference in the layout.