feat(Style): Make buttons stay pressed down
The buttons were getting darker only when there was focus on them. To test that, we could press tab and the button would not be pressed down even though it was active.
The solution was using the correct `.checked` selector.
This is an update to the changes made in the commit `a85ed0f`, when the .scss was translated into .css.