feat(PlaneSliders): Trigger re-render for play/pause, 2D view, Plane Visibility
This is a first draft of what needs to be done to get the components to re-render when pause, 2D view and plane visibility are clicked.
The objects in context do not necessarily trigger the `useSelector` to re-render if their children change. So calling `useSelector` on state.context.main is not enough.