feat(PlaneSliders/Style): Change icon image color if background is black
Since we have deleted the grayish background color for the PlaneSliders menu (commit ba5b77f), when we changed the background color it was not possible to see the Plane Visibility and the Toggle Scroll buttons.
Here we cahnge the color of such buttons to white when the background is dark.