feat(LayerEntry/ComponentSelector/Style): MUI to bootstrap, fix labelImage component option, adjust button positions and styling  (#31)
* feat(SampleDistanceSlider/Style): Align center of button with slider

 - Create different selector for icon image so we can have a rule for the right margin.  This prevents vents the overlap between the slider and the icon.
 - Introduce `margin-top` rule in the style.css to align vertically the button/image icon and the slider.

* feat(SampleDistanceSlider/Style): Simplify expression for icon-image selector

 - Delete replicate properties from '.icon-button' in style.css;
 - Define className to be both icon-button and icon-image to take advantage of the previous item

* feat(Style/GradientOpacitySlider): Adjust vertical slider appearance

Make vertical slider longer and vertically aligned with the button.

* feat(Style): Make buttons stay pressed down

The buttons were getting darker only when there was focus on them. To test that, we could press tab and the button would not be pressed down even though it was active.
The solution was using the correct `.checked` selector.
This is an update to the changes made in the commit `a85ed0f`, when the .scss was translated into .css.

* feat(BlendModeSelector): Convert from MUI to bootstrap

* feat(BlendMode/Style): Change drop-down menu minimum width

Adjust the `.blendMenu` selector property such that the words in the drop-down menu do not get chopped and fit the space.

* feat(BlendModeSelector): Swap state.context by service.machine.context wherever possible

This is an attempt to fix a merge rebase that did not carry out these changes.
It is a copy of the commit `be46e7f`.

* feat(BlendModeSelector): Bring BlendMode image icon back

By using `.icon-image` the width was set to 0 by mistake.

* feat(Style/.icon-image): Remove gray background from image icons

Use one rule for image icons with a permanent transparent background.

* feat(PlaneSliders/Style): Change icon image color if background is black (#23)

Since we have deleted the grayish background color for the PlaneSliders menu (commit ba5b77f), when we changed the background color it was not possible to see the Plane Visibility and the Toggle Scroll buttons.
Here we change the color of such buttons to white when the background is dark.

* feat(LayerEntry/ComponentSelector/Style): MUI to bootstrap, fix labelImage component option, adjust button positions and styling (#17)

* feat(SampleDistanceSlider/Style): Align center of button with slider

 - Create different selector for icon image so we can have a rule for the right margin.  This prevents vents the overlap between the slider and the icon.
 - Introduce `margin-top` rule in the style.css to align vertically the button/image icon and the slider.

*feat(LayerFiles): Convert MUI into bootstrap*

Select important state components using `useSelector` and use `service.machine` instead of `state` whenever the variable does not need to track changes;
*LayersInterface*
Replace Grid (MUI) for Col (bootstrap)
Modifications in the width and padding such that it matches the other tabs of the menu. This makes them similar to what the original version looks like.
*LayerEntry* 
	1. Delete OverlayTrigger and Button for the image icon on the right of the bar since the original is not a button.
	2.  Replace icon-button by icon-image to avoid white background when image gets clicked.
 	3. Fix the size the box and its borders to be fixed if disable the visibility button for consistency.


* feat(ComponentSelector): Fix bug to add just label without an image*

Introduce a conditional statement so we can feed a `label` to `index.html` without an `ipfsImage`.

*VolumeRenderingInputs/TransferFunction*: Prevent rendering if just have labelImage/Issue #28
Introduce new function to recognize if the input is a labelImage.
avoid ternary expression and replace it with `&&` condition;
Swap `.icon-button` by `.icon-image`