feat(ColorRangeInput/Style): Avoid warning and bug in itk-vtk viewer #346
The issue would come up when the user sets the color range to be 0 in the UI (see https://github.com/Kitware/itk-vtk-viewer/issues/346).
This version makes the color-border red when the user sets the range to be <=0. Notice, we do not SEND that information to the itk-vtk-viewer so we can avoid the issue.